Implementation rough rewrite - passes all tests There seems to be performance issues in the parallel test.
