+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/AbstractMemorySegment.java
@@ -1,0 +1,337 @@
+/*
+ * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.internal.foreign;
+
+import jdk.incubator.foreign.MemoryAddress;
+import jdk.incubator.foreign.MemorySegment;
+import jdk.incubator.foreign.SequenceLayout;
+import jdk.internal.access.JavaNioAccess;
+import jdk.internal.access.SharedSecrets;
+import jdk.internal.access.foreign.MemorySegmentProxy;
+import jdk.internal.misc.Unsafe;
+import jdk.internal.vm.annotation.ForceInline;
+
+import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Random;
+import java.util.Spliterator;
+import java.util.function.Consumer;
+
+public abstract class AbstractMemorySegment implements MemorySegment, MemorySegmentProxy {
+
+    final static int ACCESS_MASK = READ | WRITE | CLOSE | ACQUIRE;
+    final static int FIRST_RESERVED_FLAG = 1 << 16; // upper 16 bits are reserved
+    final static int SMALL = FIRST_RESERVED_FLAG;
+    final static long NONCE = new Random().nextLong();
+    final static int DEFAULT_MASK = READ | WRITE | CLOSE | ACQUIRE;
+
+    private static final Unsafe UNSAFE = Unsafe.getUnsafe();
+    private static final int BYTE_ARR_BASE = UNSAFE.arrayBaseOffset(byte[].class);
+
+    @Override
+    public final MemorySegment asSlice(long offset, long newSize) {
+        checkBounds(offset, newSize);
+        return asSliceNoCheck(offset, newSize);
+    }
+
+    @ForceInline
+    private AbstractMemorySegment asSliceNoCheck(long offset, long newSize) {
+        return new DelegatedMemorySegment(this, offset, newSize, accessModes());
+    }
+
+    @Override
+    public Spliterator<MemorySegment> spliterator(SequenceLayout sequenceLayout) {
+        checkValidState();
+        if (sequenceLayout.byteSize() != byteSize()) {
+            throw new IllegalArgumentException();
+        }
+        return new SegmentSplitter(sequenceLayout.elementLayout().byteSize(), sequenceLayout.elementCount().getAsLong(),
+                this.withAccessModes(accessModes() & ~CLOSE));
+    }
+
+    @Override
+    @ForceInline
+    public final MemoryAddress baseAddress() {
+        return new MemoryAddressImpl(this, 0);
+    }
+
+    @Override
+    public final ByteBuffer asByteBuffer() {
+        if (!isSet(READ)) {
+            throw unsupportedAccessMode(READ);
+        }
+        checkIntSize("ByteBuffer");
+        JavaNioAccess nioAccess = SharedSecrets.getJavaNioAccess();
+        ByteBuffer _bb;
+        if (base() != null) {
+            if (!(base() instanceof byte[])) {
+                throw new UnsupportedOperationException("Not an address to an heap-allocated byte array");
+            }
+            _bb = nioAccess.newHeapByteBuffer((byte[]) base(), (int)min() - AbstractMemorySegment.BYTE_ARR_BASE, (int) byteSize(), this);
+        } else {
+            _bb = nioAccess.newDirectByteBuffer(min(), (int) byteSize(), null, this);
+        }
+        if (!isSet(WRITE)) {
+            //scope is IMMUTABLE - obtain a RO byte buffer
+            _bb = _bb.asReadOnlyBuffer();
+        }
+        return _bb;
+    }
+
+    @Override
+    public final int accessModes() {
+        return accessModesInternal() & ACCESS_MASK;
+    }
+
+    abstract int accessModesInternal();
+
+    abstract AbstractMemorySegment asUnconfined();
+
+    @Override
+    public AbstractMemorySegment withAccessModes(int accessModes) {
+        checkAccessModes(accessModes);
+        if ((~accessModes() & accessModes) != 0) {
+            throw new UnsupportedOperationException("Cannot acquire more access modes");
+        }
+        return new DelegatedMemorySegment(this, 0L, byteSize(), accessModes);
+    }
+
+    @Override
+    public boolean hasAccessModes(int accessModes) {
+        checkAccessModes(accessModes);
+        return (accessModes() & accessModes) == accessModes;
+    }
+
+    private void checkAccessModes(int accessModes) {
+        if ((accessModes & ~ACCESS_MASK) != 0) {
+            throw new IllegalArgumentException("Invalid access modes");
+        }
+    }
+
+    @Override
+    public final void close() {
+        if (!isSet(CLOSE)) {
+            throw unsupportedAccessMode(CLOSE);
+        }
+        checkValidState();
+        closeNoCheck();
+    }
+
+    final AbstractMemorySegment acquire() {
+        if (Thread.currentThread() != ownerThread() && !isSet(ACQUIRE)) {
+            throw unsupportedAccessMode(ACQUIRE);
+        }
+        return acquireNoCheck();
+    }
+
+    abstract AbstractMemorySegment acquireNoCheck();
+
+    abstract void closeNoCheck();
+
+    @Override
+    public final byte[] toByteArray() {
+        checkIntSize("byte[]");
+        byte[] arr = new byte[(int)byteSize()];
+        MemorySegment arrSegment = MemorySegment.ofArray(arr);
+        MemoryAddress.copy(baseAddress(), arrSegment.baseAddress(), byteSize());
+        return arr;
+    }
+
+    boolean isSmall() {
+        return isSet(SMALL);
+    }
+
+    void checkRange(long offset, long length, boolean writeAccess) {
+        checkValidState();
+        if (writeAccess && !isSet(WRITE)) {
+            throw unsupportedAccessMode(WRITE);
+        } else if (!writeAccess && !isSet(READ)) {
+            throw unsupportedAccessMode(READ);
+        }
+        checkBounds(offset, length);
+    }
+
+    abstract long min();
+
+    abstract Object base();
+
+    private boolean isSet(int mask) {
+        return (this.accessModesInternal() & mask) != 0;
+    }
+
+    private void checkIntSize(String typeName) {
+        if (byteSize() > (Integer.MAX_VALUE - 8)) { //conservative check
+            throw new UnsupportedOperationException(String.format("Segment is too large to wrap as %s. Size: %d", typeName, byteSize()));
+        }
+    }
+
+    private void checkBounds(long offset, long length) {
+        if (isSmall()) {
+            checkBoundsSmall((int)offset, (int)length);
+        } else {
+            if (length < 0 ||
+                    offset < 0 ||
+                    offset > byteSize() - length) { // careful of overflow
+                throw outOfBoundException(offset, length);
+            }
+        }
+    }
+
+    @ForceInline
+    private void checkBoundsSmall(int offset, int length) {
+        if (length < 0 ||
+                offset < 0 ||
+                offset > (int)byteSize() - length) { // careful of overflow
+            throw outOfBoundException(offset, length);
+        }
+    }
+
+    UnsupportedOperationException unsupportedAccessMode(int expected) {
+        return new UnsupportedOperationException((String.format("Required access mode %s ; current access modes: %s",
+                modeStrings(expected).get(0), modeStrings(accessModesInternal()))));
+    }
+
+    private List<String> modeStrings(int mode) {
+        List<String> modes = new ArrayList<>();
+        if ((mode & READ) != 0) {
+            modes.add("READ");
+        }
+        if ((mode & WRITE) != 0) {
+            modes.add("WRITE");
+        }
+        if ((mode & CLOSE) != 0) {
+            modes.add("CLOSE");
+        }
+        if ((mode & ACQUIRE) != 0) {
+            modes.add("ACQUIRE");
+        }
+        return modes;
+    }
+
+    private IndexOutOfBoundsException outOfBoundException(long offset, long length) {
+        return new IndexOutOfBoundsException(String.format("Out of bound access on segment %s; new offset = %d; new length = %d",
+                        this, offset, length));
+    }
+
+    protected int id() {
+        //compute a stable and random id for this memory segment
+        return Math.abs(Objects.hash(base(), min(), NONCE));
+    }
+
+    static class SegmentSplitter implements Spliterator<MemorySegment> {
+        AbstractMemorySegment segment;
+        long elemCount;
+        final long elementSize;
+        long currentIndex;
+
+        SegmentSplitter(long elementSize, long elemCount, AbstractMemorySegment segment) {
+            this.segment = segment;
+            this.elementSize = elementSize;
+            this.elemCount = elemCount;
+        }
+
+        @Override
+        public SegmentSplitter trySplit() {
+            if (currentIndex == 0 && elemCount > 1) {
+                AbstractMemorySegment parent = segment;
+                long rem = elemCount % 2;
+                long split = elemCount / 2;
+                long lobound = split * elementSize;
+                long hibound = lobound + (rem * elementSize);
+                elemCount  = split + rem;
+                segment = parent.asSliceNoCheck(lobound, hibound);
+                return new SegmentSplitter(elementSize, split, parent.asSliceNoCheck(0, lobound));
+            } else {
+                return null;
+            }
+        }
+
+        @Override
+        public boolean tryAdvance(Consumer<? super MemorySegment> action) {
+            Objects.requireNonNull(action);
+            if (currentIndex < elemCount) {
+                AbstractMemorySegment acquired = segment.acquire();
+                try {
+                    action.accept(acquired.asSliceNoCheck(currentIndex * elementSize, elementSize));
+                } finally {
+                    acquired.closeNoCheck();
+                    currentIndex++;
+                    if (currentIndex == elemCount) {
+                        segment = null;
+                    }
+                }
+                return true;
+            } else {
+                return false;
+            }
+        }
+
+        @Override
+        public void forEachRemaining(Consumer<? super MemorySegment> action) {
+            Objects.requireNonNull(action);
+            if (currentIndex < elemCount) {
+                AbstractMemorySegment acquired = segment.acquire();
+                try {
+                    if (acquired.isSmall()) {
+                        int index = (int) currentIndex;
+                        int limit = (int) elemCount;
+                        int elemSize = (int) elementSize;
+                        for (; index < limit; index++) {
+                            action.accept(acquired.asSliceNoCheck(index * elemSize, elemSize));
+                        }
+                    } else {
+                        for (long i = currentIndex ; i < elemCount ; i++) {
+                            action.accept(acquired.asSliceNoCheck(i * elementSize, elementSize));
+                        }
+                    }
+                } finally {
+                    acquired.closeNoCheck();
+                    currentIndex = elemCount;
+                    segment = null;
+                }
+            }
+        }
+
+        @Override
+        public long estimateSize() {
+            return elemCount;
+        }
+
+        @Override
+        public int characteristics() {
+            return NONNULL | SUBSIZED | SIZED | IMMUTABLE | ORDERED;
+        }
+    }
+
+    // Object methods
+
+    @Override
+    public String toString() {
+        return "MemorySegment{ id=0x" + Long.toHexString(id()) + " limit: " + byteSize() + " }";
+    }
+}
